FROMLIST: misc: fastrpc: Expand context ID mask for DSP polling mode support
Current FastRPC message context uses a 12-bit mask where the upper
8 bits represent the context ID from idr_alloc_cyclic and the lower
4 bits represent the PD type. This layout works for normal FastRPC
calls but fails for polling mode because DSP expects a 16-bit context
with an additional async mode bit. To enable polling mode support
from DSP(DSP writes to poll memory), DSP expects a 16-bit context
where the upper 8 bits are context ID, the lower 4 bits are PD type
and the 5th bit from the end denotes async mode(not yet upstreamed).
If this bit is set, DSP disables polling. With the current design,
odd context IDs set this bit, causing DSP to skip poll memory updates.
Update the context mask to ensure a hole which won't get populated,
ensuring polling mode works as expected. This is not a bug and the
change is added to support polling mode.

Link: https://lore.kernel.org/r/20251128050534.437755-4-ekansh.gupta@oss.qualcomm.com
Signed-off-by: Ekansh Gupta <ekansh.gupta@oss.qualcomm.com>